Refiled NavItem features into smaller source files
