+++ adaxmvn/axmgc_dmo_fibo/src/main/scala/axmgc/dmo/ksrc/lean_mthlb/LeanMathlibNavRouteStuff.scala
@@ -3,28 +3,30 @@ package axmgc.dmo.ksrc.lean_mthlb
 import axmgc.dmo.fin.ontdmp.{PlainNavRouteBldr}
 import axmgc.xpr.vis_js.{NavJsonTxtApi, RobustNavJsonTxtImpl}
 
-private trait LeanMathlibFolderStuff
+private trait LeanMathlibNavRouteStuff
 
-class LeanMathlibTreeDataSrc extends RobustNavJsonTxtImpl {
+
+class LnMthLbNavJsonTxtImpl extends RobustNavJsonTxtImpl {
 	// Produces nav-item JSON responses for Lean Mathlib packages + proof-files
 
 	private val PRMKY_SEL = "sel"
 	private val PRMVL_LEAN_TREE = "leantree"
 	private val PRMVL_DUM = "dummy"
 
+	private val myLmlNim = new LmlNavItemMaker
+
 	override def mkBroadJsonTxt(paramMap: Map[String, String]): String = {
 		val selPrm_opt = paramMap.get(PRMKY_SEL)
 		val selPrm = selPrm_opt.getOrElse(PRMVL_DUM)
 		selPrm match {
-			case PRMVL_LEAN_TREE => "LEAN BROAD ANSWER" //  mkLeanMathlibNavRespBrd(paramMap)
-			case PRMVL_DUM => "FALLBACK BROAD ANSWR" // myFbackNIR.mkBroadAnswerEntity(paramMap)
+			case PRMVL_LEAN_TREE => myLmlNim.mkBigTree() // "LEAN BROAD ANSWER"
+			case PRMVL_DUM => super.mkBroadJsonTxt(paramMap) // "FALLBACK BROAD ANSWR"
 			case _ => throw new Exception("Unknown selector: " + selPrm)
 		}
-
 	}
 	override def mkNarrowJsonTxt(navQID: String, paramMap: Map[String, String]): String = ???
 }
 
 class LnMthlbNavRouteBldr extends PlainNavRouteBldr {
-	override protected def getPlainNavJsonTxtMkr: NavJsonTxtApi = new LeanMathlibTreeDataSrc
+	override protected def getPlainNavJsonTxtMkr: NavJsonTxtApi = new LnMthLbNavJsonTxtImpl
 }
